Move Projection::Jacobian to namespace
